Fixing package structures&& add/initialize websnap features
