+++ .gitignore
@@ -1,0 +1,1 @@
+./src/modules/__pycache__/
