+++ unit-testing/websnap_test/urls.txt
@@ -1,0 +1,4 @@
+https://www.google.com
+https://www.google.com/intl/he/gmail/about/for-work/
+https://www.google.com/intl/mr_in/gmail/about/
+https://www.google.com/intl/fil/gmail/about/policy/
